restructure gui startup
