+++ gui/frontend/assets.go
@@ -1,0 +1,9 @@
+//go:build !headless
+// +build !headless
+
+package assets
+
+import "embed"
+
+//go:embed all:dist
+var Assets embed.FS
