--- main_headless.go
@@ -1,14 +1,0 @@
-//go:build headless
-// +build headless
-
-package main
-
-import (
-	"fmt"
-)
-
-func init() {
-	guiFunc = func() {
-		fmt.Println("I regret to inform you, this netclient is headless")
-	}
-}
